Added better authwellknown handling
