+++ projects/angular-auth-oidc-client/src/lib/callback/refresh-session.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { forkJoin, Observable, of, throwError, TimeoutError, timer } from 'rxjs';
 import { map, mergeMap, retryWhen, switchMap, take, timeout } from 'rxjs/operators';
 import { AuthStateService } from '../auth-state/auth-state.service';
+import { AuthWellKnownService } from '../config/auth-well-known/auth-well-known.service';
 import { OpenIdConfiguration } from '../config/openid-configuration';
 import { CallbackContext } from '../flows/callback-context';
 import { FlowsDataService } from '../flows/flows-data.service';
@@ -12,7 +13,6 @@ import { LoginResponse } from '../login/login-response';
 import { StoragePersistenceService } from '../storage/storage-persistence.service';
 import { UserService } from '../user-data/user.service';
 import { FlowHelper } from '../utils/flowHelper/flow-helper.service';
-import { ConfigurationService } from './../config/config.service';
 import { RefreshSessionRefreshTokenService } from './refresh-session-refresh-token.service';
 
 export const MAX_RETRY_ATTEMPTS = 3;
@@ -24,7 +24,7 @@ export class RefreshSessionService {
     private loggerService: LoggerService,
     private silentRenewService: SilentRenewService,
     private authStateService: AuthStateService,
-    private authWellKnownService: ConfigurationService,
+    private authWellKnownService: AuthWellKnownService,
     private refreshSessionIframeService: RefreshSessionIframeService,
     private storagePersistenceService: StoragePersistenceService,
     private refreshSessionRefreshTokenService: RefreshSessionRefreshTokenService,
@@ -111,7 +111,7 @@ export class RefreshSessionService {
     allConfigs: OpenIdConfiguration[],
     extraCustomParams?: { [key: string]: string | number | boolean }
   ): Observable<boolean | CallbackContext | null> {
-    const { configId, authWellknownEndpointUrl } = config;
+    const { authWellknownEndpointUrl } = config;
 
     const isSilentRenewRunning = this.flowsDataService.isSilentRenewRunning(config);
     this.loggerService.logDebug(config, `Checking: silentRenewRunning: ${isSilentRenewRunning}`);
@@ -127,7 +127,7 @@ export class RefreshSessionService {
       return of(null);
     }
 
-    return this.authWellKnownService.getAuthWellKnownEndPoints(authWellknownEndpointUrl, config).pipe(
+    return this.authWellKnownService.queryAndStoreAuthWellKnownEndPoints(authWellknownEndpointUrl, config).pipe(
       switchMap(() => {
         this.flowsDataService.setSilentRenewRunning(config);
