+++ projects/angular-auth-oidc-client/src/lib/login/popup/popup-login.service.ts
@@ -3,7 +3,7 @@ import { Observable, of, throwError } from 'rxjs';
 import { switchMap, take } from 'rxjs/operators';
 import { AuthOptions } from '../../auth-options';
 import { CheckAuthService } from '../../check-auth.service';
-import { ConfigurationService } from '../../config/config.service';
+import { AuthWellKnownService } from '../../config/auth-well-known/auth-well-known.service';
 import { OpenIdConfiguration } from '../../config/openid-configuration';
 import { LoggerService } from '../../logging/logger.service';
 import { UrlService } from '../../utils/url/url.service';
@@ -19,7 +19,7 @@ export class PopUpLoginService {
     private loggerService: LoggerService,
     private responseTypeValidationService: ResponseTypeValidationService,
     private urlService: UrlService,
-    private authWellKnownService: ConfigurationService,
+    private authWellKnownService: AuthWellKnownService,
     private popupService: PopUpService,
     private checkAuthService: CheckAuthService
   ) {}
@@ -50,7 +50,7 @@ export class PopUpLoginService {
 
     this.loggerService.logDebug(configuration, 'BEGIN Authorize OIDC Flow with popup, no auth data');
 
-    return this.authWellKnownService.getAuthWellKnownEndPoints(authWellknownEndpointUrl, configuration).pipe(
+    return this.authWellKnownService.queryAndStoreAuthWellKnownEndPoints(authWellknownEndpointUrl, configuration).pipe(
       switchMap(() => {
         const { customParams } = authOptions || {};
